requirements.txt
tavily-python==0.3.9
neo4j==5.18.0
openai>=1.0.0
python-dotenv==1.0.1
requests==2.31.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0
reka-api>=3.0.0
pydantic>=2.0.0
